requirements.txt
langchain==0.2.14
langchain-core==0.2.32
langchain-community==0.2.12
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
faiss-cpu==1.7.4


langchain-chroma==0.1.2

chromadb==0.4.24

streamlit==1.32.0

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2

pypdf==4.1.0
python-dotenv==1.0.1

pandas
duckdb
ipykernel

 